requirements.txt
matrix-nio
ollama
pyyaml
py-daemon
matrix-nio[e2e]